Add daily_grand_repository.py, update imports and function names in daily_grand_route.py, and add daily_grand_factory.py
